added conditional to require all form fields to be filled out and added alert to confirm deleting a point of interest

+++ src/scripts/eventListeners.js
@@ -11,39 +11,49 @@ const interestId = document.querySelector("#interest_id");
 const mainContainer = document.getElementById("mainContainer")
 
 const eventListeners = {
-    addSaveButtonEventListener() {
-        // when the user clicks save after filling out the form, save the interest as an object to the database, add it to the DOM, and clear the form
-
+    addNewSaveButtonEventListener() {
         mainContainer.addEventListener("click", () => {
             if (event.target.id.startsWith("save")) {
 
-                const newInterestObject = {
-                    "placeId": placeInput.value,
-                    "name": nameInput.value,
-                    "description": descriptionInput.value,
-                    "cost": costInput.value
+                // if any of the form fields are left blank, alert user to fill out all fields
+
+                if (placeInput.value.length === 0 || nameInput.value.length === 0 || descriptionInput.value.length === 0 || costInput.value.length === 0) {
+                    alert("Please fill out all fields")
+                } else {
+
+                    // otherwise, create new object, add the interest to the DOM, and clear the form
+
+                    const newInterestObject = {
+                        "placeId": placeInput.value,
+                        "name": nameInput.value,
+                        "description": descriptionInput.value,
+                        "cost": costInput.value
+                    }
+
+                    apiManager.addNewInterest(newInterestObject)
+                        .then(() => {
+                            renderComponents.renderInterestCards(),
+                                renderComponents.clearInterestForm()
+                        })
                 }
-    
-                apiManager.addNewInterest(newInterestObject)
-                    .then(() => {
-                        renderComponents.renderInterestCards(),
-                        renderComponents.clearInterestForm()
-                    })
             }
         })
     },
-    addDeleteButtonEventListener() {
-        // delete the specific interest the user selects 
-        
+    addDeleteButtonsEventListener() {
         mainContainer.addEventListener("click", () => {
-            if(event.target.id.startsWith("delete")) {
-                const interestIdToDelete = event.target.id.split("--")[1];
+            if (event.target.id.startsWith("delete")) {
 
-                apiManager.deleteInterest(interestIdToDelete)
-                    .then(renderComponents.renderInterestCards)
+                // if a user clicks delete on an interest, alert them asking to confirm. If confirmed, delete the interest.
+
+                if (confirm("Are you sure you want to delete this point of interest?")) {
+                    const interestIdToDelete = event.target.id.split("--")[1];
+
+                    apiManager.deleteInterest(interestIdToDelete)
+                        .then(renderComponents.renderInterestCards)
+                }
             }
         })
-    }
+    },
 }
 
 export default eventListeners;
